fixed some bugs where data was getting passed and some variable name changes didn't get changed everywhere. now will zip dispite the OS using ZipArchive, no longer using exec functions to do zipping. enhanced by pulling out concatenating and zipping into their own functions so a little clearer and cleaner.

+++ app/Iep/Legacy/Handlers/Events/ZipPdfFiles.php
@@ -2,8 +2,8 @@
 
 namespace App\Iep\Legacy\Handlers\Events;
 
-use Queue;
 use \Carbon\Carbon;
+use Queue, ZipArchive;
 use App\Iep\Legacy\Pdf;
 use App\Iep\Legacy\Events\PdfWasFilled;
 use App\Iep\Legacy\Commands\RemoveFile;
@@ -12,6 +12,7 @@ use Illuminate\Contracts\Queue\ShouldQueue;
 
 class ZipPdfFiles
 {
+    public $outFile;
     /**
      * Create the event listener.
      *
@@ -33,30 +34,16 @@ class ZipPdfFiles
 		$this->queueFiles($event->files);
 
 		if (count($event->files) > 1) {
-			if ($event->fileOption == 'zip' && PHP_OS !== 'WINNT') {
-				$outfile = $event->concatName . '-' . str_random(4) . '.zip';
-				$infile = '';
-				foreach ($event->files as $file) {
-					$infile .= ' ' . escapeshellarg($file);
-				}
-
-				exec("zip $outfile $infile");
+			if ($event->fileOption == 'zip') {
+                $this->outFile = $event->concatName . '-' . str_random(4) . '.zip';
+                $this->createZip($event->files);
 			} else {
-				foreach ($event->files as $index => $file) {
-					if ($index == 0) {
-						$pdf = new Pdf($file);
-					} else {
-						$pdf->addFile($file);
-					}
-				}
-
-				$outfile = $event->concatName . '-' . str_random(4) . '.pdf';
-
-				$pdf->saveAs($outfile);
+                $this->outFile = $event->concatName . '-' . str_random(4) . '.pdf';
+                $this->concatPdfs($event->files);
 			}
 
-            $this->queueFiles([$outfile]);
-			return $outfile;
+            $this->queueFiles([$this->outFile]);
+			return $this->outFile;
 		}
 
 		return $event->files[0];
@@ -70,7 +57,44 @@ class ZipPdfFiles
 	*/
 	protected function queueFiles($files)
 	{
-		$date = Carbon::now()->addMinutes(15);
-		Queue::later($date, new RemoveFile($files));
+        $queueDriver = config('queue.default');
+
+        if (!is_null($queueDriver) && $queueDriver !== 'sync') {
+            $date = Carbon::now()->addMinutes(15);
+    		Queue::later($date, new RemoveFile($files));
+        }
 	}
+
+    /**
+     * create zip file of all pdf files
+     *
+     * @param array $files
+     */
+     protected function createZip($files) {
+         $zip = new ZipArchive();
+         $zip->open($this->outFile, ZIPARCHIVE::CREATE);
+
+         foreach ($files as $file) {
+             $zip->addFile($file);
+         }
+
+         $zip->close();
+     }
+
+     /**
+      * concat all pdfs into one
+      *
+      * @param array $files
+      */
+     protected function concatPdfs($files) {
+         foreach ($files as $index => $file) {
+             if ($index == 0) {
+                 $pdf = new Pdf($file);
+             } else {
+                 $pdf->addFile($file);
+             }
+         }
+
+         $pdf->saveAs($this->outFile);
+     }
 }
